@@ -14,16 +14,17 @@
         "bootstrap": "^4.1.0",
         "cross-env": "^5.1",
         "jquery": "^3.2",
-        "laravel-mix": "^4.0.7",
+        "laravel-mix": "^5.0.9",
         "lodash": "^4.17.13",
         "popper.js": "^1.12",
-        "resolve-url-loader": "^2.3.1",
-        "sass": "^1.15.2",
+        "resolve-url-loader": "^3.1.2",
+        "sass": "^1.29.0",
         "sass-loader": "^7.1.0",
-        "vue": "^2.5.17",
+        "vue": "^2.6.12",
         "vue-router": "^3.1.6",
         "vue-template-compiler": "^2.6.11",
         "vuejs-paginate": "^2.1.0",
         "vuex": "^3.3.0"
-    }
+    },
+    "name": "stepapp"
 }
