@@ -13,14 +13,19 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('name');
+        Schema::create('users', function (Blueprint $table) { //　usersテーブル作成
+            $table->increments('id');
+            $table->string('name')->nullable();
+            $table->unsignedInteger('group')->nullable()->default(1);
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
+            $table->string('profile',300)->nullable();
+            $table->string('image')->nullable();
             $table->rememberToken();
-            $table->timestamps();
+            $table->dateTime('created_at')->nullable();
+            $table->timestamp('updated_at')->nullable();
+            $table->boolean('delete_flg')->nullable()->default(0);
         });
     }
 
