@@ -6,7 +6,7 @@
 
   <title>@yield('title', 'ログイン')</title>
   <!-- Scripts -->
-  <script src="{{ asset('js/app.js') }}" defer></script>
+   <script src="{{ asset('js/app.js') }}" defer></script>
   <meta name="viewport" content="width=device-width,initial-scale=1">
   <!-- CSRF Token -->
   <meta name="csrf-token" content="{{ csrf_token() }}">
@@ -20,6 +20,9 @@
 </head>
 
 <body>
-  <div id="app"></div>
+
+  <div id="app">
+  </div>
+
 </body>
 </html>
